bandalists_drf/requirements.txt
appnope==0.1.0
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
backports.shutil-get-terminal-size==1.0.0
channels==0.17.2
constantly==15.1.0
coverage==4.0.3
daphne==1.3.0
decorator==4.0.6
Django==1.11.8
django-cors-headers==1.1.0
django-filter==1.1.0
django-rest-auth==0.9.0
djangorestframework==3.7.7
djangorestframework-recursive==0.1.1
easy-thumbnails==2.5
enum34==1.1.6
gnureadline==6.3.3
hyperlink==17.3.1
incremental==17.5.0
Markdown==2.6.5
msgpack-python==0.4.8
path.py==8.1.2
pathlib2==2.3.0
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5
Pygments==2.2.0
pytz==2017.3
redis==2.10.6
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
wcwidth==0.1.7
zope.interface==4.4.3
